Working on pages outside of a node project now works
The rolled up HTML generating JS now retains imports for @nakedjsx, however they are replaced with absolute paths to the installation of @nakedjsx that generated the code. This allows it work when operating as a bundled dependency of npx nakedjsx, on files that otherwise wouldn't be able to resolve '@nakedjsx/*'. This is particularly relevant to allowing it to work on page source files that live completely outside of a node project.