New 'dynamic' built-in asset import type
It's now possible to fetch JSX snippets from an arbitrary data source, and then use those to generate a series of HTML files from a single *-html.mjs file.

Crucially, it's possible for *-html.mjs to make JSX tags available to to the JSX snippet by adding them to the node built-in 'global' object.