Removed render thread pool
The performance benefit did not seem significant, and it introduce significant complexity vs a pure async model.

Further, for Page.AppendCss to be able to reserve class names to prevent generation of clashing classes, the client JS processing needs to be invoked during Page.Render(), which this change makes simpler.